mathworks removal
An attempt to remove all matlab stuff. This has led to many errors that
now seem to have nothing to do with mathworks, but just seems to remove
functionality of the cism package. so I assume I will have to come back
to this approach after I have some more knowledge of this package. so I
will commit it for now and start my own lsm with other code I have. Ill
commit it soon